+++ pyanalyze/test_generators.py
@@ -10,18 +10,18 @@ class TestGenerator(TestNameCheckVisitorBase):
     def test_generator_return(self):
         from typing import Generator
 
-        def gen(cond) -> Generator[int, str, float]:
+        def gen(cond) -> Generator[int, str, bytes]:
             x = yield 1
             assert_is_value(x, TypedValue(str))
             yield "x"  # E: incompatible_yield
             if cond:
-                return 3.0
+                return b"hello"
             else:
                 return "hello"  # E: incompatible_return_value
 
         def capybara() -> Generator[int, int, int]:
             x = yield from gen(True)  # E: incompatible_yield
-            assert_is_value(x, TypedValue(float))
+            assert_is_value(x, TypedValue(bytes))
 
             return 3