+++ pyanalyze/test_signature.py
@@ -1122,10 +1122,12 @@ class TestOverload(TestNameCheckVisitorBase):
             print("x", file="not a file")  # E: incompatible_argument
 
         def pacarana(f: float):
+            # https://github.com/python/cpython/issues/120080
+            assert isinstance(f, float)
             assert_is_value(f.__round__(), TypedValue(int))
             assert_is_value(f.__round__(None), TypedValue(int))
             f.__round__(ndigits=None)  # E: incompatible_call
-            assert_is_value(f.__round__(1), TypedValue(float))
+            assert_is_value(f.__round__(1), TypedValue(float) | TypedValue(int))
 
     @assert_passes()
     def test_runtime(self):